% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_sweep.R
\name{cluster.sweep}
\alias{cluster.sweep}
\title{Performs a hyperparameter sweep across lambda, nclust and quantile function for a dataset}
\usage{
cluster.sweep(
  X,
  coord,
  nclust = c(3, 6, 10),
  lambda = c(1),
  metrics = list(colMins = matrixStats::colMins, colMeds = matrixStats::colMedians,
    colMaxs = matrixStats::colMaxs),
  max.iter = 100
)
}
\arguments{
\item{X}{An n x p matrix with n cells and p features}

\item{coord}{An n x 2 matrix with X and Y coordinates for each cell}

\item{nclust}{A vector of numbers of clusters to sweep}

\item{lambda}{A vector of lambda values to sweep}

\item{metrics}{A named list of quantile functions to sweep}

\item{max.iter}{Maximum iterations per sweep}
}
\value{
A 'long' dataframe with cluster membership for each cell along with sweep information
}
\description{
Performs a hyperparameter sweep across lambda, nclust and quantile function for a dataset
}
