% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{k.means.spatial}
\alias{k.means.spatial}
\title{Clusters spatial omics data with spatial regularization}
\usage{
k.means.spatial(
  X,
  cell.dists,
  nclust = 3,
  lambda = 0.1,
  metric = matrixStats::colMins,
  max.iter = 100
)
}
\arguments{
\item{X}{An n x p matrix with n cells and p features}

\item{cell.dists}{An n x n matrix with pairwise distances between all cells}

\item{nclust}{A scalar indicating the number of clusters}

\item{lambda}{The weight that spatial penalty takes}

\item{metric}{A column-wise quantile function to determine distance to cluster (eg. colMins, colMedians, colMaxs from matrixStats)}

\item{max.iter}{Maximum number of iterations before stopping clustering}
}
\value{
a list containing mus (the centroid of each cluster), est (cluster membership for each cell), num_it (number of iterations taken to converge)
}
\description{
Clusters spatial omics data with spatial regularization
}
