% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_functions.R
\name{spatial.distance}
\alias{spatial.distance}
\title{Calculates the spatial distance between cells and each cluster. Distance to cluster is determined by metric.}
\usage{
spatial.distance(cell.dists, clust, metric = matrixStats::colMins)
}
\arguments{
\item{cell.dists}{An n x n matrix of pairwise physical distances between all cells}

\item{clust}{A vector length n indiacting cluster membership for each cell}

\item{metric}{A column-wise quantile function to determine distance to cluster (eg. colMins, colMedians, colMaxs from matrixStats)}
}
\value{
An nclust X n matrix with the physical distance to each cluster for each cell
}
\description{
Calculates the spatial distance between cells and each cluster. Distance to cluster is determined by metric.
}
