% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialkmeans.R
\docType{package}
\name{spatialkmeans}
\alias{spatialkmeans}
\title{Cluster Spatial Omics Data Via Spatially Regularized K-Means}
\description{
Single-cell spatial omics clustering, incorporating the physical distance between cells into the omics distance. This penalizes cells that are further away from being in the same cluster.
}
\author{
Joel Eliason \email{joelne@umich.edu}
}
